backend/app/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.9
sentence-transformers==2.2.2
numpy==1.26.0
pandas==2.2.2
python-dotenv==1.0.0
pdfminer.six==20240706
reportlab==4.0.0
aiofiles==23.2.1